Decompose src/diagnostics/checks.ts into barrel + sibling modules (warren-5bf4)
